dcu-scheduler/pyproject.toml
[tool.poetry]
name = "dcu-scheduler"
version = "0.1.2"
description = "EDEN T&S, DCU Scheduler"
authors = ["JeongYun Sim <jysim@edentns.com>"]


[tool.poetry.dependencies]
python = "^3.10"
colorlog = "^6.8.0"
pyyaml = "^6.0.1"
pydantic = "^2.6.3"
psycopg2-binary = "^2.9.9"
pytest = "^8.2.2"
peewee = "^3.17.6"
peewee-async = "^0.12.1"
aiopg = "^1.4.0"
pytz = "^2024.1"
apscheduler = "^3.10.4"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
tqdm = "^4.66.5"
aiohttp = "^3.10.6"
lxml = "^5.3.0"
beautifulsoup4 = "^4.12.2"
langchain-community = "^0.0.27"
langchain = "^0.1.11"
tzdata = "^2024.1"
langchain-openai = "^0.1.1"
pandas = "^2.2.2"
python-keycloak = "^3.9.1"
cx-oracle = "^8.3.0"
oracledb = "^1.4.2"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"